removed Grant class
Instead of returning a Grant class, where the requested resource can be accessed via the "resource" attribute, the resource itself is returned. The initial idea of the grant class was to also provide the user information alongside, but this is in an API not that useful in contrast to a classical web app.